Roughly document renderer and group camera logic and buffers together
Took 37 minutes
