--- crates/dream-renderer/src/camera_uniform.rs
@@ -1,23 +1,0 @@
-use nalgebra::Matrix4;
-
-use crate::camera;
-
-#[repr(C)]
-#[derive(Debug, Copy, Clone, bytemuck::Pod, bytemuck::Zeroable)]
-pub struct CameraUniform {
-    view_proj: [[f32; 4]; 4],
-}
-
-impl CameraUniform {
-    pub fn update_view_proj(&mut self, camera: &camera::Camera) {
-        self.view_proj = camera.build_view_projection_matrix().into();
-    }
-}
-
-impl Default for CameraUniform {
-    fn default() -> Self {
-        Self {
-            view_proj: Matrix4::identity().into(),
-        }
-    }
-}
